+++ src/core-atoms/delim.ts
@@ -37,7 +37,11 @@ export class MiddleDelimAtom extends Atom {
 
   _serialize(options: ToLatexOptions): string {
     if (
-      !(options.expandMacro || options.skipStyles) &&
+      !(
+        options.expandMacro ||
+        options.skipStyles ||
+        options.skipPlaceholders
+      ) &&
       typeof this.verbatimLatex === 'string'
     )
       return this.verbatimLatex;
@@ -90,7 +94,11 @@ export class SizedDelimAtom extends Atom {
 
   _serialize(options: ToLatexOptions): string {
     if (
-      !(options.expandMacro || options.skipStyles) &&
+      !(
+        options.expandMacro ||
+        options.skipStyles ||
+        options.skipPlaceholders
+      ) &&
       typeof this.verbatimLatex === 'string'
     )
       return this.verbatimLatex;