+++ src/core-atoms/enclose.ts
@@ -107,7 +107,11 @@ export class EncloseAtom extends Atom {
 
   _serialize(options: ToLatexOptions): string {
     if (
-      !(options.expandMacro || options.skipStyles) &&
+      !(
+        options.expandMacro ||
+        options.skipStyles ||
+        options.skipPlaceholders
+      ) &&
       typeof this.verbatimLatex === 'string'
     )
       return this.verbatimLatex;