+++ src/core-atoms/extensible-symbol.ts
@@ -104,7 +104,11 @@ export class ExtensibleSymbolAtom extends Atom {
 
   _serialize(options: ToLatexOptions): string {
     if (
-      !(options.expandMacro || options.skipStyles) &&
+      !(
+        options.expandMacro ||
+        options.skipStyles ||
+        options.skipPlaceholders
+      ) &&
       typeof this.verbatimLatex === 'string'
     )
       return this.verbatimLatex;