+++ src/core-atoms/group.ts
@@ -41,7 +41,11 @@ export class GroupAtom extends Atom {
 
   _serialize(options: ToLatexOptions): string {
     if (
-      !(options.expandMacro || options.skipStyles) &&
+      !(
+        options.expandMacro ||
+        options.skipStyles ||
+        options.skipPlaceholders
+      ) &&
       typeof this.verbatimLatex === 'string'
     )
       return this.verbatimLatex;