+++ src/public/mathfield.ts
@@ -9,6 +9,7 @@ import { Selector } from './commands';
 | `"latex"`             | LaTeX rendering of the content, with LaTeX macros not expanded. |
 | `"latex-expanded"`    | All macros are recursively expanded to their definition. |
 | `"latex-unstyled"`    | Styling (background color, color) is ignored |
+| `"latex-without-placeholders"`    | Replace `\placeholder` commands with their body |
 | `"math-json"`         | A MathJSON abstract syntax tree, as an object literal formated as a JSON string. Note: you must import the CortexJS Compute Engine to obtain a result. |
 | `"math-ml"`           | A string of MathML markup. |
 | `"spoken"`            | Spoken text rendering, using the default format defined in config, which could be either text or SSML markup. |
@@ -28,6 +29,7 @@ export type OutputFormat =
   | 'latex'
   | 'latex-expanded'
   | 'latex-unstyled'
+  | 'latex-without-placeholders'
   | 'math-json'
   | 'math-ml'
   | 'spoken'